[RHI] render target pooling (#43)
* [RHI] render target pooling

Add RHIRenderTargetPool: render targets (with pool-owned backing images)
are keyed by their full attribute set and reused across frames and
renderer recreations instead of being reallocated.

- a target returns to the pool automatically when all external
  references to it and its images are dropped; it becomes reusable after
  the frames-in-flight safety delay, or immediately after a device-idle
  flush (e.g. on pipeline switch)
- free targets unused for RetireAfterFrames frames are released
- GPU/Forward/Deferred renderers now acquire their single-color targets
  (scene color, tone mapping / screen color) from the pool, so switching
  pipelines reuses matching images
- new render_target_pool test case verifies distinct targets while held,
  reuse of a freed target, and retirement; runs in CI on the Windows
  lavapipe runner after the functional test

* [RHI] fix -Wshorten-64-to-32 in RHIResourceRef::use_count on Apple builds

std::shared_ptr::use_count() returns long; returning it as int only
triggers the warning once the template member is instantiated, which
the render target pool's freeness check now does.

* [Test] fix lambda capture shadowing in RenderTargetPoolTest

The init-capture named 'task' shadows the RunOnRenderThread parameter,
rejected by -Wshadow-uncaptured-local on the Windows clang build.

* [RHI] replace render target pool auto-retirement with a manual ReleaseUnused API

The pool no longer ages out free targets after a frame-count heuristic;
they stay pooled until RHIRenderTargetPool::ReleaseUnused() is called.
Deletion still goes through the deferred-deletion path, so the call is
safe while frames are in flight. The render_target_pool test's third
stage now exercises the manual release instead of waiting out the
retirement window (which also shortens the test).

* [Test] cover pooled render target reuse across runtime pipeline switches

+++ libraries/include/rhi/RHIResource.h
@@ -184,7 +184,7 @@ public:
         return ptr_.get();
     }
 
-    [[nodiscard]] int use_count() const // NOLINT
+    [[nodiscard]] long use_count() const // NOLINT
     {
         return ptr_.use_count();
     }
